DISCARD: Add ezchip and arceb-ezchip basic machine type
This commit needs to be submitted separately to <config-patches@gnu.org>
and then back-ported into the GCC tree.

This allows for users to specify configure targets like 'arceb-ezchip'
or even just 'ezchip'.  Currently, despite having this patch in the tree
we don't make use of this feature, and could safely remove this patch.

My concern with the patch as it currently stands is that it locks the
relationship of ezchip means arc.  This is fine now, but it's always
possible that at some future point this might not be true.  I think we
might consider the following courses of action for this patch:

 1. Drop it.  We don't currently need this functionality, we always
 configure with a full target quad (arceb-ezchip-linux-uclibc) which
 works fine with config.sub as is, so lets leave this problem for
 another day.

 2. Go ahead with this as is.  If there's ever a non-arc ezchip product
 then that will need to use a more specific descriptor, as 'ezchip' will
 be reserved for the arc product line.

 3. Pick a new descriptor instead of 'ezchip' and rework this patch.  We
 could include the product name maybe.

ChangeLog:

	* config.sub: Add ezchip and arceb-ezchip as acceptable
	$basic_machine.