GLOBAL: gcc: Some addresses can be considered constant
Treat the addresses of global and static variables as constant when they
are bound locally.

The thinking is that this change allows additional optimisation for
EZchip, though there is a concern that this change may have negative
side effects for other targets.  We need to carry out additional testing
of this commit, or restructure it to be arc specific in some way.

gcc/ChangeLog:

	* builtins.c (fold_builtin_constant_p): Consider addresses for
	global/static variables as constant when they are bound locally.

gcc/testsuite/Change Log:

	* gcc.c-torture/execute/20130418-1.c: New test.
	* gcc.c-torture/execute/20130418-2.c: Likewise.
	* gcc.c-torture/execute/20130418-3.c: Likewise.
	* gcc.c-torture/execute/20130418-4.c: Likewise.
	* gcc.c-torture/execute/20130418-5.c: Likewise.
	* gcc.c-torture/execute/bcp-1.c (global): Initialise.
	(bad10): Moved to...
	(good3): ...here.