DISCARD: Disable use of --gc-sections for building libs
This works around a bug in binutils, where we report that --gc-sections
works, but in reality it does not.  This needs fixing in binutils, and
this patch needs to be dropped.

Actually, need to check if this is still an issue at all with current
binutils, this might have been fixed.