ARC: set JUMP_TABLES_IN_TEXT_SECTION
When code runs from section loaded into fast memory
we do not want that it will use rodata section from slow memory.
Now switch jump table will always be inlined in same code section
where from switch is called.

Later we should do it automatically but only when code section is
not the default code section (.text) that we assume is not loaded to
fast memory.

Signed-off-by: Noam Camus <noamc@ezchip.com>