arc/gcc: Unconditional annulled branches in arc_predicate_delay_insns
The function 'arc_predicate_delay_insns' currently applies some criteria
to find jumps with delay slots, and then asserts that all of the jump
instructions it has found will be conditional if-then-else style jumps.

However, it only makes this assertion for branches that are annulled,
which seems like a sensible thing to do; if a branch is annulled then
surely the branch must one that might be taken, or might not be taken.

The problem is that I found a branch that was marked annulled, but was
unconditional, which feels like a bug.

Before I fully understood the bug, my initial approach was allow that
annulled branches might NOT be conditional, this is the patch you see
here.

However, I now believe that this is the WRONG approach to solving this
problem, better would be to enforce the rule that only conditional
branches should be marked as annulled.

To make life extra annoying, the bug I was chasing ended up being masked
by a different commit, 522e1febb5d3109a.

In the test case I had found an epilogue instruction is placed into the
delay slot of an unconditional branch, and the branch is then marked as
annulled, for some reason.  The commit 522e1febb5d3109a prevents such
epilogue instructions (RTX_FRAME_RELATED_P) being placed into the delay
slot of an annulled branch, and so the bug disappears.

I still suspect that in the right circumstance we could end up with a
non-epilogue instruction being placed into the delay slot of a
non-conditional branch, and the branch then being marked as annulled,
which would re-trigger this bug.  However, for now I plan to park this
patch, and return to it when I have more time.

gcc/ChangeLog:

	* config/arc/arc.c (arc_predicate_delay_insns): Don't assume all
	jump instructions are conditional.

gcc/testsuite/ChangeLog:

	* gcc.target/arc/predicate-delay.c: New file.