arc/gcc: Replace rI constraint with r & Cm2 for ld insns
In some of the load instructions the constraint 'rI' was being used,
which would accept either a register or a signed 12 bit constant.  The
problem is that the 32-bit form of the ld instruction only takes a
signed 9-bit immediate.  As such, some ld instructions would be
generated that would, when assembled be 64-bit long, however, GCC
believed them to be 32-bit long.  This error in the length would cause
problems during branch shortening.

gcc/ChangeLog:

	* config/arc/arc.md (*loadqi_update): Replace use of 'rI'
	constraint with separate 'r' and 'Cm2' constraints.
	(*load_zeroextendqisi_update): Likewise.
	(*load_signextendqisi_update): Likewise.
	(*loadhi_update): Likewise.
	(*load_zeroextendhisi_update): Likewise.
	(*load_signextendhisi_update): Likewise.
	(*loadsi_update): Likewise.
	(*loadsf_update): Likewise.