DISCARD: arc/gcc: Turn on LTO FAT objects by default
This patch works around the fact that we are currently building the
nps400 toolchain using a modern GCC but with a somewhat old version of
binutils.

The version of binutils that we are using does not support slim lto
objects, which is what modern GCC generates by default.  The work around
for this is to turn on fat lto objects by default.

This patch will not be pushed upstream, when we rebase the nps400
binutils changes this patch will become redundant and can be safely
dropped.

Until then the slim lto objects can still be generated using
-fno-fat-lto-objects, however, this will likely break in many cases
unless you are using a recent version of binutils-gdb.

gcc/ChangeLog:

	* common.opt (ffat-lto-objects): Make this flag default on.