ezchip/gcc: Update expected results on movb-1 test
The expected results on this test were wrong.  I checked against
comments in the assembler, and by running a functional test on the
simulator, GCC appears to be producing the correct instruction, and the
test was checking for the wrong thing.

This commit updates the expected test results.

gcc/testsuite/ChangeLog:

	* gcc.target/arc/movb-1.c: Update expected results.