GLOBAL: gcc: Allow code partitioning in more cases
Currently, placing a variable into a custom section results in the
reorder-blocks-and-partition feature being deactivated.  This seems
unnecessary, and the comments on the variable
user_defined_section_attribute (the mechanism by which the feature is
deactivated) suggests the deactivation should only have happened when a
function is placed into a custom section.

This commit renames user_defined_section_attribute to
func_with_user_defined_section_attribute to make its intended use a
little clearer, and then ensures that the flag is only set when it is a
function that has a custom section applied to it.

This change will aid targets that place data into custom sections, but
leave code in the default section, and still wish to perform code
partitioning.

gcc/ChangeLog:

	* bb-reorder.c: Update comment on 'toplev.h' include.
	(pass_partition_blocks::gate): Update
	user_defined_section_attribute to
	func_with_user_defined_section_attribute.
	* c-family/c-common.c (handle_section_attribute): Update
	user_defined_section_attribute to
	func_with_user_defined_section_attribute.
	* final.c: Likewise.
	* toplev.c: Likewise.
	* toplev.h: Likewise.