gcc: Don't pass nps400 cpu type to the assembler
In preparation for using the 6.2 GCC branch with the old (2.23 based)
assembler, this commit stops GCC passing the nps400 cpu type to the
assembler, and instead uses the arc700 cpu type.  The old assembler only
added all nps400 instructions into the arc700 architecture, and did not
have a separate assembler flag for nps400.

gcc/ChangeLog:

	* config/arc/arc.h (ASM_SPEC): Use arc700 as the assembler cpu
	when assembling nps400 code.
	* config/arc/arc.c (arc_init): Use ARC700 as the cpu string
	embeded into the assembler file for nps400 code.