gcc/nps: Move __NPS400__ definition out of CPP_SPEC
The __NPS400__ definition is taken care of as part of the arc-c.def
mechanism.  The additional definitions in the CPP_SPEC are not needed
and can cause confusion, they only appear when -mcpu=nps400 is
mentioned on the command line, while the definition from arc-c.def
correctly appears in both the case where -mcpu=nps400 is used, and in
the case where GCC is configured with --with-cpu=nps400.

gcc/ChangeLog:

	* config/arc/arc.h (CPP_SPEC): Remove unneeded __NPS400__
	definition.