gcc: Increase insn length for doloop_begin_i in some cases
The length for some variants of doloop_begin_i failed to take into
account the possible alignment that might be added.  As a result, in
some cases the doloop_begin_i would be bigger than expected, this causes
problems in branch shortening.

As Synopsys have a rewrite of zero overhead loops (ZOL) in progress
there seems little point investing a large effort improving the length
calculation of loops for now, instead I just increase the length to take
into account the maximum doloop_begin_i size.

The downside, is that in a few edge cases we might generate long branch
instruction when a short one would have done.  For the short term I
think this is acceptable, when Synopsys' new ZOL implementation lands
then we should get much better length computation, and so we should get
the short branches back.  However, backporting a new ZOL implementation
to the 6.2 branch is probably rather risky at this stage.

gcc/ChangeLog:

	* config/arc/arc.md (doloop_begin_i): Increase length of some loop
	variants.