gcc/riscv: Fix passing C++ structs with empty sub-structs
Empty structs have size 1-byte in C++.  The RISC-V ABI document[1]
says:

   Empty structs or union arguments or return values are ignored by C
   compilers which support them as a non-standard extension. This is
   not the case for C++, which requires them to be sized types.

Which I interpret to mean that the 1-byte sized empty struct should be
passed like any other argument, which I believe should result in the
single byte being passed as an integer argument.

Currently when we pass something like this:

  struct sf { struct {} e; float f; };

GCC get this completely wrong, it tries to ignore 'e' and then passes
the bytes covered by 'e' as the contents for 'f'.  However, if we pass
this:

  struct si { struct {} e; int f; };

In this case GCC does pass this as a two integer struct (on RV32) with
one integer covering 'e' and one covering 'f'.

This patch updates GCC to correctly pass C++ empty structs.  The 'sf'
case will now be passed as an integer and a float, while the 'si' case
will still be passed as two floats.

[1] https://github.com/riscv/riscv-elf-psabi-doc/blob/master/riscv-elf.md

gcc/ChangeLog:

	* config/riscv/riscv.c (riscv_flatten_aggregate_field): Identify
	empty structs with non-zero size, and record these as fields that
	need passing.
	(riscv_pass_aggregate_in_fpr_and_gpr_p): Count structs seen at
	this point as integer arguments.

gcc/testsuite/ChangeLog:

	* g++.target/riscv/call-with-empty-struct-float.C: New file.
	* g++.target/riscv/call-with-empty-struct-int.C: New file.
	* g++.target/riscv/call-with-empty-struct.H: New file.
	* g++.target/riscv/riscv.exp: New file.