WIP: Ignore empty structs when passing C++ structs
This replaces the previous patch with a new approach.  Now empty C++
structures are ignored completely, so passing this:

    struct sf { struct {} e; float f; };

Will be passed as a single float argument covering 'f'.

Currently this patch does not handle this case yet:

   struct si { struct {} e; int f; };

Here, GCC still passes this as two integer arguments, however, this
should be easy enough to fix, however, this would be ABI breaking.

I've raised an issue with the ABI document to discuss the right way
forward[1] and depending on the results there will depend the right
approach moving forward.

[1] https://github.com/riscv/riscv-elf-psabi-doc/issues/88