[wip] clean up
Remove this code block.  Most of this was compiled out anyway, and the
remaining code was debug only.  However, the use of DECL_STATIC_CHAIN
was problematic, as this is only valid to call on a FUNCTION_DECL, and
some of the trees that this function receives are not pure
FUNCTION_DECLs, but could be INDIRECT_REF to ADDR_EXPR to
FUNCTION_DECL.

Rather than rewrite this code to handle that case, just so I can print
some debug (which would have to be removed eventually anyway), just
delete the code now.